object-oriented-pretrained/requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
easydict==1.13
filelock==3.29.0
fsspec==2026.3.0
gdown==6.0.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.12
Jinja2==3.1.6
kornia==0.8.2
kornia_rs==0.1.10
llvmlite==0.47.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.1
pillow==12.2.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
PyMatting==1.1.15
PySocks==1.7.1
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scipy==1.17.1
setuptools==81.0.0
shellingham==1.5.4
simsimd==6.5.16
soupsieve==2.8.3
stringzilla==4.6.0
sympy==1.14.0
timm==1.0.26
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transparent-background==1.3.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wget==3.2
